sdk extraction > cleanup, add java doc
