+++ src/main/java/com/hivemq/edge/modules/adapters/data/DataPoint.java
@@ -7,7 +7,13 @@ import com.hivemq.extension.sdk.api.annotations.NotNull;
  * Instances of it can be created via {@link com.hivemq.edge.modules.adapters.factories.DataPointFactory} accessible via {@link com.hivemq.edge.modules.adapters.factories.AdapterFactories} in the {@link com.hivemq.edge.modules.adapters.model.ProtocolAdapterInput}.
  */
 public interface DataPoint {
+    /**
+     * @return the value of the data point.
+     */
     @NotNull Object getTagValue();
 
+    /**
+     * @return the tag of the data point.
+     */
     @NotNull String getTagName();
 }
