+++ src/main/java/com/hivemq/edge/modules/adapters/data/ProtocolAdapterDataSample.java
@@ -12,16 +12,35 @@ import java.util.List;
  *
  */
 public interface ProtocolAdapterDataSample {
+    /**
+     * @return the {@link AdapterSubscription} containing information how the data gets published by the broker.
+     */
     @JsonIgnore
     @NotNull AdapterSubscription getSubscription();
 
+    /**
+     * @return the timestamp when this data sample was taken.
+     */
     @JsonIgnore
     @NotNull Long getTimestamp();
 
+    /**
+     * Adds a new data point to this sample.
+     *
+     * @param tagName the name for this data point.
+     * @param tagValue the value for this data point.
+     */
     void addDataPoint(@NotNull String tagName, @NotNull Object tagValue);
 
+    /**
+     * Sets/Overwrites all data points of the sample with the given argument.
+     * @param list the new list of data points for this sample.
+     */
     void setDataPoints(@NotNull List<DataPoint> list);
 
+    /**
+     * @return the list of data points
+     */
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @NotNull List<DataPoint> getDataPoints();
 }
